% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleSearchR.R
\name{concat_cols}
\alias{concat_cols}
\title{A function to concatenate two columns with a user-defined separator}
\usage{
concat_cols(df, col1, col2, sep = "+")
}
\arguments{
\item{df}{the data frame containing the columns to concatenate}

\item{col1}{the name of the first column to concatenate}

\item{col2}{the name of the second column to concatenate}

\item{sep}{the separator to use between the two columns (default is "+")}
}
\value{
the concatenated column as a character vector
}
\description{
A function to concatenate two columns with a user-defined separator
}
\examples{
df <- data.frame(
  Organization_Name = c("Walmart", "First Republic Bank", "Netflix", "Amazon", "Snap", "Comcast", "Apple"),
  Headquarters_Location = c("Bentonville, Arkansas, United States", "San Francisco, California, United States", "Los Gatos, California, United States", "Seattle, Washington, United States", "Venice, California, United States", "Philadelphia, Pennsylvania, United States", "Cupertino, California, United States")
)

concat_cols(df, "Organization_Name", "Headquarters_Location", sep = " + ")
}
